Factor `distribute_plan()` function into distributed_query_planner.rs
